12003 ios updates

+++ ChartIQDemo/ChartIQDemo/StudyDetailViewController.swift
@@ -150,7 +150,7 @@ class StudyDetailViewController: UITableViewController {
     func keyboardWillShow() {
         colorPicker.isHidden = true
     }
-    
+
     func updateStudyParameters() {
         var inputs = [String: Any]()
         for input in inputParameter {
@@ -243,14 +243,14 @@ class StudyDetailViewController: UITableViewController {
             cell.labels?[1].text = (parameter["value"] as? String)?.capitalized ?? ""
         case .number, .text:
             cell.textFields![0].keyboardType = optionType == .number ? .numberPad : .default
-            cell.textFields?[0].text = optionType == .number ? String(parameter["value"] as? Int ?? 0) : parameter["value"] as? String ?? ""
+            cell.textFields?[0].text = optionType == .number ? String(parameter["value"] as? Float ?? 0) : parameter["value"] as? String ?? ""
             cell.textFieldValueDidEndEditingBlock = {[weak self] (cell, textField) in
                 guard let strongSelf = self else { return }
                 let name = cell.labels![0].text ?? ""
                 for (index, var inputs) in strongSelf.inputParameter.enumerated() {
                     if inputs["name"] as? String == name {
                         if optionType == .number {
-                            inputs["value"] = Int(textField.text ?? "0")
+                            inputs["value"] = Float(textField.text ?? "0")
                         } else {
                             inputs["value"] = textField.text ?? ""
                         }
